MicroBitLog.cpp & MicroBitLog.h: testing

+++ inc/MicroBitLog.h
@@ -423,7 +423,7 @@ namespace codal
         /**
         * New function to facilitate reading logged data from the uBit directly
         */
-        ManagedString MicroBitLog::readRow(uint8_t rowIndex);
+        ManagedString MicroBitLog::readRow(uint32_t rowIndex);
     };
 }
